move current_token_pending_claims to token_v2 processor
